LayoutTests/imported/w3c/web-platform-tests/html/browsers/browsing-the-web/back-forward-cache/service-worker-clients-claim.https-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: SyntaxError: Invalid URL

Harness Error (TIMEOUT), message = null

TIMEOUT Clients.claim() evicts pages that would be affected from BFCache Test timed out
